% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_outliers.R
\name{flag_outliers}
\alias{flag_outliers}
\title{Highlight variables containing certain threshold of outliers}
\usage{
flag_outliers(df, threshold = 0.1)
}
\arguments{
\item{df}{A dataframe or tibble containing numeric variables}

\item{threshold}{Threshold (minimum) required to flag variables with outliers}
}
\value{
Tibble containing name of variables and proportion of outliers contained
}
\description{
Highlight variables containing certain threshold of outliers
}
\examples{
df <- data.frame('col1'=c(1,2,3), 'col2'=c(1,2,3))
flag_outliers(df, threshold=0.2)


}
