% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correlated_features.R
\name{get_correlated_features}
\alias{get_correlated_features}
\title{Calculate correlation between all feature pairs in the numeric dataframe.
Returns feature pairs having correlation higher than the threshold value.}
\usage{
get_correlated_features(x, threshold, consider_sign = FALSE)
}
\arguments{
\item{x}{Input DataFrame or tibble containing numeric variables}

\item{threshold}{Float value for correlation above
which feature pairs will be returned}

\item{consider_sign}{(boolean) determines whether correlation value has to
be checked for magnitude only or for sign
(positive/ negative) also. Default checks only the magnitude.}
}
\value{
DataFrame Tibble containing containing feature1, feature2,
and corresponding correlation.
}
\description{
Calculate correlation between all feature pairs in the numeric dataframe.
Returns feature pairs having correlation higher than the threshold value.
}
\examples{
x <- data.frame('age'=c(23, 13, 7, 45),
                'height'=c(1.65, 1.23, 0.96, 1.55),
                'income'=c(20, 120, 120, 25))
get_correlated_features(x, threshold=0.7)

}
