% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_imputer.R
\name{missing_imputer}
\alias{missing_imputer}
\title{Return the imputed version of data based on the methods selected}
\usage{
missing_imputer(data, method = "mean")
}
\arguments{
\item{data}{The raw data with missing values}

\item{method}{The method used to impute the missing values}
}
\value{
An imputed dataframe with the method specified
}
\description{
Return the imputed version of data based on the methods selected
}
\examples{
data <- data.frame('col1'=c(1,NA,3), 'col2'=c(1,2,3))
missing_imputer(data , method = "median")


}
