% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{overview}
\alias{overview}
\title{Gives a statistical overview of the input data,
returns a \code{data.frame} of descriptive statistical values.}
\usage{
overview(input, quiet = FALSE)
}
\arguments{
\item{input}{A dataframe or tibble containing numeric data}

\item{quiet}{Optional boolean value for creating global variables
for use elsewhere in the environment, default is FALSE}
}
\value{
Data.frame containing all calculated values
}
\description{
Gives a statistical overview of the input data,
returns a \code{data.frame} of descriptive statistical values.
}
\examples{
df <- data.frame('col1'=c(1,2,3), 'col2'=c(2,3,6))
overview(df, quiet=FALSE)

}
